Move icon/convection/ to convection/icon/
Convert 7 files relative imports to absolute, update test files,
icon_physics.py, and parameters.py.
