+++ jcm/physics/convection/icon/tracer_transport.py
@@ -11,7 +11,7 @@ from typing import NamedTuple, Tuple, Optional
 
 from .updraft import UpdatedraftState
 from .downdraft import DowndraftState
-from ..constants.physical_constants import tmelt
+from jcm.physics.icon.constants.physical_constants import tmelt
 
 
 class TracerIndices(NamedTuple):
