+++ jcm/physics/icon/icon_physics.py
@@ -22,7 +22,7 @@ from jcm.physics.icon.icon_coords import IconCoords
 # Import physics modules (will be implemented progressively)
 from jcm.physics.icon.radiation.radiation_scheme import radiation_scheme
 from jcm.physics.icon.icon_physics_data import RadiationData
-from jcm.physics.icon.convection import tiedtke_nordeng_convection
+from jcm.physics.convection.icon import tiedtke_nordeng_convection
 from jcm.physics.icon.clouds import shallow_cloud_scheme, cloud_microphysics
 from jcm.physics.icon.parameters import Parameters
 from jcm.physics.icon.surface import surface_physics_step, initialize_surface_state
